arm: DT: Tianchi: Correct color swap property
That one wasn't really working in the previous implementation,
so that was a never-noticed bug we had in there.

Tianchi needs no color swap, so just set it as RGB just for
future reference, even if setting it like that is basically a
NOP, because it is the default setting.