msm: ipa_v2: Fix ipa_trace.h header location
ipa_trace.h is in drivers/platform/msm/ipa/ipa_v2/.
The previous code was ok if we wanted to compile the kernel with
the Android build system, because that one is copying all the
required headers in a specific location.

We want to be able to compile the kernel normally, without using
a very heavy build system, but just the good ol' kernel build system.