clk: msm: clock-cpu-8994: Update __smp_call_function_single usage
Coming from 3.10, __smp_call_function_single was changed here.

The new function is renamed to smp_call_function_single_async
as the difference with smp_call_function_single is that the call
is asynchronous and can be done from contexts with disabled
interrupts.
Also, the wait parameter got removed.
Update the source accordingly.