arm64: spinlock: Add errata for explicit SEV in spinlock unlock
In certain unexplained cases, the stlr alone might not wakeup
the processor waiting in WFE on a spinlock.
Add an explicity dsb and SEV in write_unlock, read_unlock
and spin_unlock to ensure that the core waiting on the lock
wakes up from WFE.
This is activated by default for MSM8994.