gpu: msm: KGSL: Add legacy stack, version LA.HB.1.3.1_rb1.15
This is the worst hack ever seen in the story of Linux kernel.

This also is a temporary commit: we need to keep compatibility
with older QC userspace for Adreno until we complete the migration
from kernel 3.10 to 3.18.

The reason for this commit is that new 3.18 KGSL kernel driver
uses a new way to manage Adreno ring buffers, which is implemented
in userspace libraries from the new UM branches.
We first want to bringup the boards and be able to boot android
and then, eventually, try the new QC userspace because we are
not sure that older (3xx, 4xx) Adreno chips are even supported
on there.