arm: DT: MSM8994: Do not map dfps_data_mem as DMA pool
We have bootloader informations in that memory range, describing
the DFPS for MDSS PLL.
Mapping it as shared DMA pool is simply not right: it will zero
out the memory space and we will lose the informations stored in
there, hence we will be unable to setup MDSS PLL like bootloader
wanted us to.

Leave the memory descriptor for MDSS PLL driver to get the
memory range, but DO NOT map it as any DMA pool.