defconfig: loire: disable LEGACY_PTYS
A pseudo terminal (PTY) is a software device consisting of two
halves: a master and a slave. The slave device behaves identical to
a physical terminal; the master device is used by a process to
read data from and write data to the slave, thereby emulating a
terminal. Typical programs for the master side are telnet servers
and xterms.

Linux has traditionally used the BSD-like names /dev/ptyxx
for masters and /dev/ttyxx for slaves of pseudo
terminals. This scheme has a number of problems, including
security.  This option enables these legacy devices; on most
systems, it is safe to say N.

Signed-off-by: Alin Jerpelea <alin.jerpelea@sonymobile.com>