defconfig: loire: enable LOCALVERSION
Append an extra string to the end of your kernel version.
This will show up when you type uname, for example.
The string you set here will be appended after the contents of
any files with a filename matching localversion* in your
object and source tree, in that order.  Your total string can
be a maximum of 64 characters.

Signed-off-by: Alin Jerpelea <alin.jerpelea@sonymobile.com>