media: msm: camera_v2_legacy: Optimize code style for 3.18 CAF
The NEWGEN camera driver has some code style optimizations.
We are currently porting some commits to avoid too many
modifications to the msmb_isp header and to try to unify the
newgen and legacy header as much as possible for userspace to
build and work correctly, and for this reason we need the
functions in there (and whole camera api) to be as similar
as possible to the newgen driver.
Easy commit porting ftw.