fbdev: msm: mdss_smmu: Add variable init check to avoid panic
The kernel can panic in there. And it's very easy.
The mdss_mdp driver calls clock initialization from the probe
function by calling the footswitch control for continuous splash:
this *sometimes* requires starting clocks, either due to a bad
configuration coming from bootloader, or due to a bad clocks
handling at init time from the MDSS driver.

Either way, the mdss_mdp_probe function initializes mdp mdata,
which is REQUIRED for the mdss_smmu to initialize all the structs
and all the variables (mdss_smmu dynamically sets those to support
various IOMMUs), hence when we make a call to get clocks on we
actually also try to attach the IOMMU.

While this is a strict requirement in basically all conditions,
when booting with continuous splash on, the IOMMU is already
configured by the bootloader, hence we can skip this configuration
safely.

When we are not in continuous splash mode, the clocks will not be
onlined in footswitch control, so will be fine as earlier.