iommu: qcom_iommu: Add secure partition table calls for downstream SCM
The driver was using equivalents from the mainline qcom_scm driver.
We unfortunately are on downstream drivers, hence, write functions
to use our current driver for SCM calls.
While at it, add a check for SCM service version: this is mandatory
if we want to also support the latest SoCs using IOMMUv2, as on
those ones we need to set the CP pool size prior doing the secure
partition table initialization.