fbdev: msm: somc_panel: Read PA HW configuration preventively
The MDSS API changed on new k4.4. This new API supports multiple
hardware versions configurable with PA v2 config struct.

Preventively read the PA HW configuration to get the current
version of the PA in the MDSS hardware macroblock, so that we
are able to feed it back to the MDSS API for it to choose the
correct functions to update the parameters on the HW.