defconfig: loire: disable IKCONFIG
Kernel .config support

This option enables the complete Linux kernel .config file
contents to be saved in the kernel. It provides documentation
of which kernel options are used in a running kernel or in an
on-disk kernel.  This information can be extracted from the kernel
image file with the script scripts/extract-ikconfig and used as
input to rebuild the current kernel or to build another kernel.
It can also be extracted from a running kernel by reading
/proc/config.gz if enabled (below).

Signed-off-by: Alin Jerpelea <alin.jerpelea@sonymobile.com>