usb: gadget: ci13xxx_udc: Ensure spin unlock in ISR
If for some strange reason we get a NULL ci13xxx structure,
the function was just returning without doing anything.
The problem here is... the spinlock was getting unlocked
only after the NULL pointer check, resulting in a possible
deadlock when the USB driver gets in this bad condition.

Move the spin unlock earlier so that we avoid getting any
stupid behavior.