usb: gadget: ci13xxx_udc: Add multiple td support to hardware{en,de}queue
This patch removes the restriction of having a limited amount of only
four active tds on one endpoint. We use the linked list implementation
to manage all tds which get added and removed by hardware_{en,de}queue.

The removal of this restriction adds the driver to run into a hardware
errata. It's possible that the hardware will still address an transfer
descriptor that already got cleaned up. To solve this the patch also
postpone the cleanup of processed tds by one.

Commit ported from upstream chipidea driver
2e270412968d961ecde347343ffa67dfe39f6c95