mmc: sdio: Force the lowest voltage on card init for bcm
We don't want to give a choice to the kernel. We already know
which voltages can we use and, got to the mmc_select_low_voltage
function, we are already checking if the 165-195 OCR is available.

If this is available, just force it as this is the only voltage
range that we support (card wants 1.8V).