supply: qcom: qpnp-fg-gen3: Set profile loaded status when SoMC EXT
When using SoMC extensions we load the battery profile in a
different way and it seems that we don't need to program it
in the FG3 SRAM, or not in the profile load worker for some
reason.

After loading the battery profile, there is a sort of sanity
check that checks the VBATT prediction and current voltage:
at this point, if there is no error, FG SOC is ready and we
can safely signal that the battery profile has been loaded.