arm64: DT: arm-smmu-8998: Bring up MMSS NoC AXI clock for MMSS SMMU
Access to the MMSS SMMU needs the MMSS NoC AXI to be up: we
have always been able to access it because we were making
sure that this clock got enabled before access from the users
of it.
To make it proper, add the clock dependency to the SMMU itself.