Android.mk: Add rules to generate DTBO image
This allows to generate a DTBO image from the kernel's Android.mk
for kernel inline building on AOSP: if called, it will by default
generate a dtbo image named "dtbo-arch.img" where "arch" is the
target architecture.
For example, on ARM64, the image will be named "dtbo-arm64.img".

It is possible to override the image name by setting the env
variable called "BOARD_DTBO_IMAGE_NAME" in your BoardConfig.mk:
this will generate a file named "dtbo-myname.img" in the out
folder.

Once image generation is done, it will be copied with the
standard name that the Android build system wants "dtbo.img"
directly into the out folder again.
This flexibility may be removed if in the future the build system
will change to add it by default.

This code is inspired to Christian (myself5) and Luca(luca020400)'s
kernel.mk changes for inline kernel building with dtbo.