clk: qcom: clk-hfpll: Properly check for regmap_read retval in wait loop
The return value for regmap_read was incorrectly interpreted:
this function returns zero if success, or standard error codes
if failure occurs.

Useless to say that - if failure occurs - in that read, it is
unrecoverable from that function so, instead of losing more
than 500 precious microseconds looping around the nothing, just
immediately return failure so that a recovery mechanism can
kick in.

This also solves an implicit bug that would incorrectly judge
the PLL as "LOCKED" when regmap_read fails more than once.