clk: qcom: dsi-pll-10nm: Set cached rate during dsi_pll_enable
When we try to set a rate on the VCO that is equal to the cached
rate, the clock framework will bail out and return success
without actually calling set_rate on the clock.

This would be good - if it wasn't for the fact that the PLL will
lose its clock rate configuration (and other registers) whenever
we shut down the GDSC for it (so, whenever we turn the PLL off):
to overcome to this issue, force setting the cached and expected
rate on the clock that we are trying to enable, so that we don't
fall in unclocked access issues and/or PLL enable failures.